Changed how synteny styles were being set. Previously child adapters were setting the style to their own style, not the synteny style (in AbstractApolloAdapter.clearOldData). Then a second change of style occurred in DataLoader, and this is where the synteny style would get set. In the case of regular apollo there would just be 2 redundant settings of the style.
Now the synteny adapter sets the style of its children adapters with
the synteny style, overriding the default style of the child from
apollo.cfg. So when the child adapter loads in synteny, it now loads
the synteny style file. I took out the style loading in data loader as
this makes it redundant, and the style has to be set in the child
dataadapter before load or we get tier exceptions, but we dont need 2
settings of the style.

When we get the ability to switch between single and multi species
(cant do this at the moment) these styles will need to be swapped when
switching modes. Just a note for the future.