Jalview can now display all features, whether they have a real alignment or not. Previously it only displayed features that had real alignments (as well as exons).
Now if a feature doesnt actually have
an alignment it will still display in jalview with what i call a
trivial alignment, which is just using the feats sequence (ungapped of
course) for both the ref and hit sequence.

This was per Len Triggs request. Now you can see how gene predictions
and what not align with real alignables which can be handy.

Jalview uses Alignables, and the whole point of Alignable is for
jalview to consume. It was going to be rather awkward to have jalview
deal with both Alignables and seq features that were not alignables -
it would get to the point of why have alignables. So my feeling was
that is was nice to keep to the notion that everything passed to
jalview is an alignable.

However I did not make SeqFeature implement Alignable. Instead I made
an inner class in SeqFeature called TrivialAlignable that implements
Alignable, to separate out the alignable part of seq feature from a
regular seq feature. For this it was handy to separate Alignable from
SeqFeature. Alignable only needs just a few methods from
SeqFeature. An Alignable only needs the methods that are needed for
jalview to do its alignment, it does not need the zillions of methods
that come along with SeqFeatureI (or even RangeI - though it would be
nice to break out just the high low start end stuff in rangei).

FeaturePair first sees if it has a real alignment, and if not it
queries its SeqFeature super class for the trivial alignment.

Alignables can longer be gotten from seq features
by doing instanceof and casting - you
have to do hasAlignable, getAlignable because SeqFeature has an
Alignable but it is not an Alignable itself. I suppose FeaturePair
could follow suit.

Exon no longer implements Alignable as its now covered by SeqFeatures
trivial alignment - took out alignable methods.

Added Alignable methods to FeatureList.

Len Trigg sent an example that jalview doesnt deal with properly -
gotta deal with that - i think its because there are features that are
3 bases long that dont come up in our data.