I set out to make site view hidden in synteny, which I did.  Its initial state is configurable (InitialStateForStartStopCodonVisibility "false") as well as changeable from view menu with a new view menu item. Its a bit of a real estate hog and synteny is strapped for real estate. Even if this is true, it still only appears at proper zoom level, if false doesnt appear at any zoom level.
But then I got into fiddling with all the view stuff as i found things
to be a bit quirky/buggy.

I changed reset view so it resets the view back to the initial state
specified in the style.

ShowAnnotations and ShowResults (and
InitialStateForStartStopCodonVisibility) are all just initial states
for the view menu items now, used to be permanent and the menu items
had no effect. So the user can bring up annot view or result view if
initially hidded, and vice versa. (rename ShowAnnotationsInitially?
ShowResultsInitially?)

On load with style change (game->gff) all views go to state specified
in style.
On load with no style change views keep state user has set.

Also I fixed view menu updating. In some cases they were not updating
properly. ViewMenu now queries szap for view state before coming
up. This is easier than the StrandVisibilityListener stuff, so I got
rid of the stradvisListener stuff. This fixes synteny's view menu
state bug also, where it would get all wrong switching from one
species to the other - it would not update for the newly selected
species.

ViewMenu is accessing szap directly for this view stuff. It could use
a ViewVisibilityI interface or something, that szap implements or has,
just to encapsulate and such.

StrandVisibilityListener could be cvs removed as its no longer
used. But theres gonna be visibility issues with synteny so it might
get reactivated and modified. Dont know yet.

szap.setSelectedStrandVisible is now setSelectedViewVisible and it
makes sure both strand and view is visible. Currently it assumes
non-annots are results, and site view is not included. If this is a
bummer can ammend it.

Im still not happy with synteny menus. I dont like having to make a
species active to fiddle with it. We're going to make species active
on clicking its szap, that will be an improvement. I think either the
synteny menu should be right next to the view menu or even better the
species panel active menu items should be part of the view menu
itself. Even better it would be nice if the popup menus included a
bunch of the view items at the bottom (probably only for synteny not
for single-species apollo). Then the view items would work on the szap
clicked on. It should at least go in scale menu, but it could go in
result and annot menus as well. (if it was in link menu it would have
to operate on both species i guess)

CurationSet - made nameAdapter transient as its not serializable and
it was causing exceptions on autosave. If we need this as part of
backup then we need to name adapters serializable.