For as long as the DefaultNameAdapter has been in existence, it has labelled the FeatureSet using the name of the first feature in the set. This is really odd, right? For instance, if an emsembl-generated  'Transcript' (FeatureSet) has a set of SeqFeatures(exons) then the DisplayName of the Transcript's drawable should be the ensembl Transcript Id...not the ensembl id of its first exon.
So - if the name of FeatureSet is anything except 'NO_NAME' then I'm going to use
it. Otherwise, it will look into its elements for their first name.

It works nicely over ensj (we used to hack the result we wanted by setting all the
exons in the transcript to have the transcript name). Now waiting for problems from
other places...