Fixed peptide-not-updating bug (when curator changes a transcript model, peptide wasn't getting changed). There is now a peptideValidity (should this be peptideValid?) flag that can be set to false, thus triggering retranslation when the peptide sequence of the Transcript is requested.  Otherwise, the existing translation is used (if there isn't one, then of course it is translated).
Also added a method (which we may end up removing) that sets the cDNA and
peptide IDs, as well as all the exon IDs.  (This only happens when asked--just changing the transcript ID does not
trigger this method.)  This can be called by the annot info editor if the
user changes the transcript ID.

Had to restore controversial peptide-naming call that calls Config.getNameAdapter().
Suzi feels that this is evil because Config lives in gui, and we shouldn't be
calling gui code from datamodel code.  I see her point, but I maintain that
NameAdapter is not really a gui thing, and it just lives in gui.Config because
there wasn't a better place to put it.
I am open to constructive suggestions about how else to name peptides.  (The
way it was, they were getting named -RA like the transcript, which is *wrong*.)