well, this is a big commit. waited a long time to proceed with it.
1. holder no longer exists as an inherent aspect of seqfeatures. this
is all handled now within the gui drawables. each drawable has a couple
attributes that determine when to draw the feature it is associated
with. one is the draw-level. this indicates the depth of the feature
hierarchy that the drawable can represent. e.g. a draw_level of 2
means it can draw a seqfeature and its direct parent and a draw_level
of 1 means it can only represent the seqfeature itself.

2. drawables no longer implement the datamodel interfaces. they are
truly just drawables/glyphs that can represent the model.

3. added 2 more analysis parsers for repeatmasker and trnascan

4. added some gui utilities to reduce duplicated code

5. changed the format of the tiers files to be tag : value. the code
will still read the old format, but will automatically update it to
the new format.

hmmm, think that is it. still doing more testing, but so far so good