changed the default for setting the height of the rectangles according to the score to false. otherwise things that are not specified in the tiers files come up very skinny--so skinny they are virtually invisible.
there are also a couple of little problems.
1. if you change the type of annotation using the pull-down menu in the
annotation editor there is no guarantee that these types (indicated in
the style files) have corresponding entries in the tiers files. apollo
will quietly create new default feature properties for them, but this
seems like it should be an error. Perhaps the annotation types should
only be specified in the tiers file so that it is just in one place
(needs to go on the to-do list)

2. If you do change the type of annotation it may look like it has dissapeared
entirely. In fact it is there, but because apollo does not automatically
set the vertical scroll to the drawn position you must go searching for it.