Change-in-progress to switch back&forth between synteny & normal views without reloading. Actually you're now looking at a 'synteny' view all the time, it just happens that sometimes you're only viewing one species. I have knowingly 'slammed' two classes that Suzi altered two days ago: don't panic ! Here's why:
(1) In apollo/gui/LoadUtil Suzi removed commented out code (firing region change events)
-- I have also removed that code, but altered the method name from 'addToCurationSet'
(which is misleading I think) to 'fireRegionChangeEvents' which is what it does...

(2) CompositeApolloFrame. Suzi has removed all the commented out code in this one
I just did the same, and then added in my changes.

-- so it all appears to work, except that:
-- There's a nullpointer happening when you shut apollo down. I'm looking into that.
-- Also the Controller problem is still there (current fix is hacky). I think Mark's idea
of a single controller for everything is fine, and I'll try that one first...