Fixed this exception: trying to bring up jalview on bright green result AF125387 in annot-types.xml.  maybe you could look into that? java.lang.ArrayIndexOutOfBoundsException: 1160         at jalview.datamodel.FeatureBlock.jalviewToGenomic(FeatureBlock.java:512)         at jalview.datamodel.Mapper.condensed2genomic(Mapper.java:70)         at jalview.datamodel.CondensedAlignment.createSequences(CondensedAlignment.java:224)         at jalview.gui.AlignFrame.<init>(AlignFrame.java:95)         at apollo.gui.menus.AlignMenuItems.putUpJalview(AlignMenuItems.java:118)         at apollo.gui.menus.AlignMenuItems.alignRegion(AlignMenuItems.java:136)
it turned into a bit of a slippery slope. There were 2 problems here,
a hack and bad data. AF125387 has a hit alignment but no query
alignment, so its missing a crucial indel which causes jalview to get
off and access beyond the array above. It was doing this in figuring
the end genomic coordinate, but figuring the end genomic coordinate
was a hack where it was just using the length of finished aligned
string, which was rather bad as it if the feature ended before the
alignment it would pad out, and the genomic end would get numbered with all
the padding. So I fixed the start end labeling (a bug I knew about but
was forgoing and figured it was time to fix it). I added a new class
BlockAlignment to preserve the genomic start and end of the aligned
sequences (used to get tossed) for this purpose. But theres still the
issue of what to do with the bad data - ill send mail around about this.