This is a sizable commit, but still needs further debugging.
The biggest changes are the removal of GenericAnnotations, which
have been replaced by AnnotatedFeatures. These do what Mark has
long advocated: the FeatureSet is at the top and SeqFeatures
inherit from them. Didn't do it for everything, only annotated
features. This means that an Exon is now a 'set', but one that
can never have child features. I tried to find most of the bugs
due to set methods being called instead of the more appropriate
non-set methods, but there are likely still to be more lurking.
Be on the look-out for problems that might be called due to this
change. All-in-all I like this new way of doing things and think
we should consider it for all of the features. Another thing to
look out for is misinterpreted casting, since Gene as a class
is no longer there it takes actual code to determine if the
AnnotatedFeature is at the top or is a Transcript/Exon.

In addition fixed 2 small other bugs.
1. the annotation tree panel is up again
2. <ctrl>c and <ctrl>e no longer throw array out of bounds exception