fixed my own bug, where labels were being used for annotation types instead of (correctly) using the annotation type (aka resulttype)
insertion sites are now inherited by the annotations so that they
are correctly positioned.

finally add new code so that annotations will automatically be of
the correct type if this is indicate by the type of evidence that
is used to generate the annotation. I.e., if a tRNA is pulled
down then the annotation is automatically a tRNA and so forth.
This needs to be specified in the tiers file