Fixes to get jbuilder to compile - classes not in the makefile had compile errors.
Removed DrawableSpliceSite as its not used anymore.

Commented out constructor in TranslationView - this is not used - is
this a work in progress or pase?

DrawableDisplayTool - not used by code - is this a debugging tool that
someone uses - if not used anymore we should just delete it.

I propose that SeqFeatureI.getFeatures be renamed getChildren() for
clarity sake. getFeatures seems vague to me. also getFeatureAt(int i)
rename to getChild(int i).