in scaleview ignore repaint if curation set is null, because that means data is still being loaded
fixed call to set name so that it occurs after the transcripts
have been added

do a wee-bit to speed up paint if there are loads of selections
in scaleview. if this isn't enough then we will have to go back
to the old way where these weren't shown at all until you were
more zoomed in.