Set out to fix this bug:
Transaction records aren't saved in backup.

which i did. AnnotationChangeLog now gets serialized and read back by
SerializedDiskAdapter, so transactions are preserved across backups as
they should be.

AnnotationChangeLog was being created and retrieved in Config - but
AnnotationChangeLog has nothing to do with Config. Config has become a
dumping ground for what basically are singletons.
So I took AnnotationChangeLog out of Config and made it a singleton,
which is a trivial change. Also AnnotationChangeLog can deserialize
itself from an ObjectInputStream and set its singleton instance.

Controller should likewise be taken out of Config and made a
singleton. It makes sense for a class to manage its instance.

AnnotationChangeLog and all the other editing stuff dont belong in
apollo.gui, I think they should be repackaged into apollo.editor.
config stuff into apollo.config.

AnnotationEditor.purgeTranscript was both sending out an
AnnotationChangeEvent and directly accessing AnnotationChangeLogs
vector of deleted transcripts and adding to
it. AnnotationChangeLog.handleAnnotationChangeEvent should have been
doing the latter but it wasnt, it was just noting that the gene
changed. So I changed handleAnnChange to actually add the deleted
transcript to the vector, and got rid of the direct access to the
vector.

on deleting a multi exon transcript, every exon fires an annotation
change event - cant we just do 1 for the transcript?

In setting the controller for AnnotationChangeLog i found several
different controllers lying around. First it was using the one Config
makes (should remove that one), then I ammended that to use the one in
setController that comes from the AF constructor, but that wasnt jibin
either because CAF.initialiaseApolloFrames was creating a new
Controller with every load (which i think it shouldnt) and stuffing it
in AFs instanceVar hash which sets the controller via introspection
which bypasses the setController method.

Also since ApolloFrame was using a pase controller regionChangeEvents
and lazySequenceEvents were not registering. For regionChangeEvents
this didnt seem to make much of a difference - evidence panel was
being reset by apollo panel, and the invalidate/validate didnt seem to
be needed (take this method out). The lack of lazy seq events caused
the lazy seq popup to not come up - thats now
fixed. AbstractLazySequence now gets its Controller from Config on
every access (change this to get from Controller singleton method).

The fix for now is that initialiseApolloPanels is calling
setController with the new Controller its making.

But the whole controller thing needs to be refactored:

Controller should be made a singleton (not a static in Config).
One Controller should be made and reused - not this remaking on every
load.
There should be one controller for all of the species - not one for each.
(also need to refactor instance hash as we all know)