new (private) method  getIDPrefix returns the prefix of the appropriate ID format (e.g. for a gene would return "CG"). This is so when we're evaluating the legality of an ID, if it has "temp" or "tmp" in it we can just look at the prefix and not reject the ID for not (say) ending with numbers.
Renamed some methods for clarity, e.g. getId -> makeNewId (because
that's what it does--it doesn't get an existing ID; it makes a new
one).

Removed snarky (and obsolete) comment--what's the point of sticking
snarky comments in where the intended recipient might or might not
happen to come across them?

Don't make names or IDs for exons.  Exons are now anonymous.