conf/game.internal.style should be copied to conf/game.style in order to use it.  (Meanwhile, the old default game.style is moved to game.public.style.) The copying is automatically done by "make private" which I use to make  the internal release, so you shouldn't need to do anything special. Don't copy game.internal.style  to ~/.apollo/game.style, because Apollo automatically parses first conf/game.style and then ~/.apollo/game.style, so this will lead to redundancy.  All that should go in  ~/.apollo/game.style is any preferences you have that DIFFER from  game.internal.style.
Again, users shouldn't need to worry about any of this; "make private"
will handle it.  I just thought it would be less confusing for (internal)
users for game.style to be the style file that is used, and have it
import game.public.style, rather than having game.internal.style used,
because you might not think to look at that.