Fixed TypePanel mem leak. New type panels were being created in CAF for every species on every load and hooked up to the species controller, which is also created for every species on every load, and the controller was leading to the old data set. But presently there is only one tiers/properites used for all species, and only one TypePanel gets displayed.
So now TypePanel is a singleton, and only one is ever created. It is
wiped out on style changes (not on region change with style the same),
but not recreated, and its for all species.

It needs to listen to all species controllers, so MultiController is
being used as the master controller. More work needs to be done with
controllers but this is a start to have something for
TypePanel. Controller.getMasterController is the access for the
MultiController.

More stuff will be needed with controllers and type panel when we want
to implement separate type panels for each species (as sima has
requested), but we are not there yet (synteny Style doesnt do this) so
this will work for now.

Theres more leaks to be had but this is a start.