Ah - i can finally commit! sourceforge has been down all morning. A few things:
Fixed clas cast exception in game data adapter. When data loaded from
url (eg gene query) and edited, and then new data asked for, and save
dialog comes up and answer yes - a curation set was going to game
adapter which was casting it to a composite data holder. It now casts
to apollodataI and queries that for cdh or cs. My feeling is GA should
always receive cur set not cdh.

Also fixed null pointer in ScaleView. Nearly impossible to repeat
since its a thread thing (but sima got it to happen) but the cur set
can be nulled out mid draw. It now checks for null on all cur sets. In
the future even better would be to disable drawing during a
load(DataListener/RegionChange). I comtemplated locking out set cur
set til draw is done but this slows down load for an unneeded draw.

Fixed compile error in TranscriptEditPanel where it was accessing
superclass prive - put in protected accessor method.