for translational frame shift the drawing and the save and reading should be working
same as above for readthrough of stop codon

some questions remain on frame shift to see if final aa is present

selection is now quite different. the selected item's source
only indicates where the original select originated from. it
does not imply anything about what views may contain representations
of that selection. there is now a method in the views to return
the sub-selection of the overall selection representing what
features that view draws (inverts how this was previously done)

removed some unused files.

merged my sequenceedit class and the jalview one. michele and
i independently created classes that were identical. very cute
and gratifying to discover.

selectfeature should now automatically scroll

add a new annotation change event REPLACE. this should be
useful when we get to implimenting undo. now in the annotation
text editor if the changes are commited the cloned version
simply replaces the older original version.

the edits to account for sequencing errors on the genomic is
still in the works.