fix 1) It seems like I can't save my frame shifted genes properly. If I try to save the "amn" gene with its frame shift and new start codon, then reopen the file, I get the new start codon but I lose the frame shift. See my file:
and
2) The frame shift is not getting the translated codons right. If in
the "amn" gene I put a -1 frame shift at location 19618382 (base 2639 of
exon1 mRNA), it should slip back one base pair and read the TTT as "F".
But instead it is reading the TTT as "I".

The first was because the frameshift position can NOT be set until
the exons are added and the translation start if first set. So
adapter has to hang on to them until this has been done and then
set them.

the second bug was because when you get a single base none of the
rev-comping is done. solved by getting 2 bases and chopping it to 1