fixed the exception that was being thrown when transcripts were deleted in exon editor.
along the way also fixed another bug for duplicating transcripts and
keeping the selection when results were/are flipped.

in general there is a tension/finesse that is needed to deal with editing
changes and the impact that these have on selection changes. deletes
obviously need to be removed from the selection, but the selection should
not be cleared. at the same time the changes to the selection that occurs
as a result should not generated feature selection events until the
entire edit is completed because this causes a bunch of confusion. That is,
the drawing (from the feature select event) must happen -after- it is
notified of the model changes.