Fixed this bug: Synteny GFF example--all red features come up labeled no_name (used to have ENST names) (Vivek says this is happening with a lot of the EnsJ features, too)
In fact this was happening with all non-game(ensj,gff,gff-synteny...)
DrawableGeneFeatureSets. DrawableGeneFeatureSet was overriding getDisplayLabel
where it didnt used to and the overriding was causing the no_name. Getting rid
of the override restored the label for non-game, and game is still fine, so its
unclear why the override was there - but maybe im missing something.
I think it might've come from the merging of different drawables.

Also made EvidenceSelectionSet and TranslationView compilable for jbuilders
sake - i think they should both be deleted as they arent used anymore -
is anyone attached to them?