fixing this bug: If you select two transcripts of different annotations and choose "Merge transcripts" a window pops up asking for confirmation for merging overlapping genes, which in theory is fine.  However, if you choose "Don't merge" or close the window or as far as I can tell perform any sort of action the transcripts merge.
In other words, there is no way to prevent the merge of the transcripts if you
decide that you don't want to do so after this window has popped up.


if you ask to move exons to transcript, or
merge transcripts, and the resulting operation would leave one gene with
no more exons, you get asked something like:
There are overlapping genes in this region: guf and CG8862
If you move the selected exons, these genes will merge.
Do you want to merge these overlapping genes into a single gene?

if you say "Don't merge", then it does nothing and just returns--if
you're not prepared to merge the genes, then you can't merge the
transcripts and leave an empty gene.  is that ok?