Got game synteny working for syntenic blocks between gadfly3 and gadflydpsef1. It still needs a lot of work but the basics are working. GameSynteny was previously hardwired for dpse724 to gadflyi. That hardwiring has been ripped out. it is now hardwired for syntenic blocks (3 to dpsef1) but i started to work in more generic stuff with an eye towards being able to configure in arbitrary links - as this is the next step. So theres a new LinkConfig object that holds configuration for a link. Its currently an inner class of GameSynteny, but soon will be an outer class (i want to put it in a new apollo.conf package).
I needed to query curation sets for all of seq feats of a given
analysis type, so i added a method to curation set that does as much
returning a stranded feature set. and i added getTierFeatures to
SeqFeatureUtil which takes a feature property(analoguos to a visual
type)  and returns all the leaves of that type/property - as the
configging of the links happens at the (visual) type level not the
analysis type level as there currently isnt anyway to configure for an
analysis type in the tiers file, and its not really needed at this
point for synteny.

Added a new tiers file synteny-game.tiers which includes new types
from gadfly3 and gadflydpsef1 (im certainly open to having the colors altered).