got loading a second species from a "syntenic block" link working. syntenic block go both ways (dpse to mel and mel to dpse) so what species/database to use is figured in LinkConfig given the species of the selected link. syntenic blocks link by location. the location i get from the syntenic block is url ready(colon delimited) - the game adapter was expecting a space delimited location that it then parses and was choking on the colon delimited - it now checks what kind of location it was passed.
links are not loaded in LoadUtil at the moment (maybe eventually they
should - i forget the reason now that i circumvented LU) so no region
change event was firing and synteny menu was not getting put in on
going from 1 to 2 species(the lack of species menu was causing a null
ptr on selection). fixed this so region change fires and synteny menu
inserts on 2nd species load.

some data adapter exceptions were getting caught and then a new one
was constructed and fired - its unclear why this was so (kind of odd) but its
annoying for debugging as you lose the source of the stack
trace. changed it to pass on the original exception.