Working on bug where all syntenic blocks are selecting on links selection. This is now fixed for link species feature selection - still needs to be fixed for link panel polygon selection. The problem was the names of syntenic block features were just the chromosome or contig id (thats just how the data is). synteny link panel was doing a selection by name - and all the names were matching so everything was selecting. The reason its doing name selection is that for ensj and gff the only way to connect to linked features is by name - the links are loaded separately so at load time we dont know which links go with what features - and theres no need to know until somethings selected.
game is different - the links dont come from a separate source. The
links come from the species data itself. Therefore we know the
features that are linked at load time because we derived the links
from these features themselves. So with the new Link object we can
record the linked species features (previously FeaturePairs were used
which had no way of recording the linked feats and the info was lost
and had to do name linking). So at selection time it now queries if we
actually have the linked feats in the link and if so (true for game)
it now does selection by feature (which of course is also more
efficient than doing a name lookup). Default is to link-select by name (ensj
and gff).

Suzi had a good suggestion of making Selection a subclass of
FeatureList which i did - turned out not to be so hard and it makes
negotiating feature lists and selections a lot easier. Selection is
really just a specialization of a feature list - its a list of
features with some extra selectiony stuff. also selection can now use
feature lists functions and some of the feature listy funstions in
selection have migrated to feature list where they really belong.

some parts of the code are still very vectory (AnnotationEditor) so i
put in a Selection.getSelectedVector and held off on changing all the vectors.

SeqFeature.getFeatures should eventually return a feature list - i
held off from that for now.