first checkin on new branch
I have created a branch in the cv module called bdgp-dev in order to
introduce some changes to the cv module

These changes should be backwards compatible - they involve additional
stuff only. All queries and read-only middleware will not be adversely
affected. Loading software would have to be modified to take advantage of
the new columns.

I made this change on a branch so as not to disrupt software development
or the forthcoming releases. I would like to get these changes
incorporated into the main branch in one of the forthcoming releases, but
I am not familiar enough with all the pieces of chado middleware to know
if this will cause too much churn.

The summary is here:

-- Changes in branch:
-- new table: cvtermprop                  FOR FUTURE USE
-- new column cvterm.is_obsolete          LOADERS WILL REQUIRE CHANGING
-- new column cvterm.is_relationshiptype  LOADERS WILL REQUIRE CHANGING
-- new column cvterm_dbxref.is_for_definition    NO IMMEDIATE CHANGES
REQUIRED
-- new column cvtermsynonym.type_id              NO IMMEDIATE CHANGES
REQUIRED
-- insert statement to ensure is_a relationship is always present