Fixed the peptide coordinate bug in game synteny. The blastx hit coords were in peptide coords which i didnt realize and i was treating them like nucleotide coords throwing everything off. This was a pretty crucial bug to fix for the release. Added getGenomicPosForPeptidePos to FeatSet and SeqFeat - if we keep adding peptide stuff as i think we might we may want a peptide object.
So Im commiting this on the release_1_4_1_patches branch that i
started this morning while it was unclear what we were gonna do. As it
turns out we are probably going to continue with the main branch for
the public release (the bug fix is a pretty important one that should
go in the release - nomis testing it now). However, i think
this is a good test of a potentially good practice. I think if we do
little release branches right before the release, then it makes it
easier for the developers to continue on with the main branch while
the release stuff is being tested and all and not getting disturbed by
developments for the next release. So the changes here are pretty
small and in theory it should be easy to merge these in with the main
branch in the next day or 2.