i thought it would be cool to show the names of the reverse-strand annots in red.  this turned out to be tricky.  the only way i found to set the color of a tree node is by putting html code into the name of the seq feature in the node (there is not a DefaultMutableTreeNode.setColor method, for example).  so i did this:
        for (int i = sorted_set.size()-1; i >= 0 ; i--) {
          SeqFeatureI sf2 = (SeqFeatureI) sorted_set.elementAt(i);
          // Make reverse-strand annot text red
          if (sf2.getStrand() == -1) {
            sf2 = sf2.cloneFeature();   // gotta clone first so we don't change the name of the REAL annotation object
            sf2.setName("<HTML><FONT COLOR=RED>" + sf2.getName() + "</FONT></HTML>");
          }
          makeTree(sf2, newnode);
        }

which worked, sort of: the reverse strand annotation names indeed show up
in red, as do its transcripts.  however, when i clicked on a transcript
name to show the exons, the exon names looked like:

exon <HTML><FONT COLOR=RED>CG12173-PA</FONT></HTML> (frame 3): 1234-9875

(because the <HTML> tag isn't the first thing in the name, it doesn't get
interpreted).

i tried changing the toString method in Exon to print the transcript name
before the type ("exon"), but then the exon names come out as just the
transcript names--i guess when the interpreter gets to </HTML> it
finishes off the string and doesn't include the plain-text stuff after
it.

so then i fixed that by writing a new stripHTML method and modifying
Exon's toString method:
    String parentName = apollo.util.SeqFeatureUtil.stripHTML(getRefFeature().toString());

which worked!  but was maybe kind of a lot of effort to go to just for
red rev-strand annots.  but, hey, it looks cool.  (the html-stripping
step should be harmless for normal transcript names; it will only affect
the red ones in the annot tree).