Fixed vertical scrolling bug in results. When a synteny link was being selected the results were not being properly vertically scrolled - which made synteny link selection rather lame. It turns out the issue was analysis types that only had features on one strand. An empty analysis feature set was still being created on the other strand (a dummy if you will) and a drawable was created for this. In FeatureTierManager.synchDrawablesWithTiers a features parent's tier is set (this allows for genes to be scrolled to). The drawables parent for the dummy feat set was the very top of the drawable heirarchy - so the dummy would set the tops tier index, and the top would set all of its descendants tier index which was all drawables on that strand.
So I took out the dummy analysis feature set. So no dummy drawable is
created and this fixes the vertical scroll problem - at least the one
that i was seeing - there could be other vertical scroll bugs im missing.

The main issue with this
is that a feature being moved to the opposite strand where there is no
opposite strand feats of that type fails. This was actually buggy
before now its buggy in a different way - im gonna look into how hard
this will be to ammend - its a somewhat obscure bug.