fixed this one:  > >4. In annot info editor, if you try to change annot symbol or ID, no  > >matter what you do you get "Error: annotation symbol foo is inconsistent  > >with ID CG1234.  You must fix this problem before changes can be  > >committed."  > >[3/22 This is still happening to me, even though the idformats have now  > >been removed from game.tiers.]  > >It did let me change the name to a CG matching the ID, but only after I  > >fixed all the -PA etc. transcript suffixes.
This was independent of my changes. Someone on 2/14 changed checkIDFormat
to always return return if no format was specified. This is fine, but...
this does not mean that the Name and ID share a common format which was
a downstream consequence of this change. Now, if no format is provided,
this returns false.