Links are finally configurable in game synteny! This was a long time coming. Very exciting. Links are configured in the tiers file like this:
synteny_link_type : PEPTIDE
link_query_species : Drosophila pseudoobscura
link_hit_species : Drosophila melanogaster


valid link types are PEPTIDE and SELF. PEPTIDE is a link to a peptide
annotation (translated transcript) like a blastx. SELF (im open to
other naming suggestions here) links to features of the same type,
like syntenic blocks. TRANSCRIPT is on the list to add - shouldnt be
too hard.

PropertyScheme can be queried for all of its link types.

Species are mapped to databases in the style file:
SpeciesToDatabase "Drosophila melanogaster -> gadfly3"

This is for loading of links. Theres a bug here and i think i know
what it is - gonna fis this right away.

In putting all the stuff needed for a link in FeatureProperty, it
suddenly became unnecasary to replicate that in a LinkConfig object,
so I got rid of LinkConfig and I expanded the Linker objects, which
now do most of the linking work - AbstractLinker, PeptideLinker,
SelfLinker and LinkerI. GameSynteny mainly loads the appropriate
linker and doles out the loading of the links to it.

PeptideLinker's inner class ResultTranscriptPair deals with the
one-way direction of that link.