added ability to parse BLAST output that has the topcombo option
also fixed a bug where setRange (in AbstractSequence) was silently
setting the length of the sequence to be the same as the range.
This is BAD. I deleted this line. Anyone who is under the impression
that the length of the range and the length of the sequence are
equivalent is mistaken. Any assumptions in their code that are
based on this mistake will result in problems.