So there was request to have style selection be at the database level. In other words to (optionally) be able to load a different style for each database in the dataadapters database list. So I have implemented this, but I feel a little funny about it. So the way this is specified is in the style file with:
DatabaseToStyle "dbname -> dbstylefile.style"

So what feels funny is that the style one specifies in apollo.cfg for
the dataadapter may no longer be the style that gets loaded for that
adapter if it has db styles within it. I guess it feels funny for a
style to be able to load other styles. It would be fine if it loaded
different tiers, as thats what styles do. But users would like to be
able to change style stuff as well per database apparently.
Im open to suggestions here.