Added in Johnathan Crabtrees TIGR chado adapter. Its up to date with the latest apollo and compiles and everything. To actually use it you need a jdbc sybdriver and you would have to uncomment the sybdriver code that is commented out to get it to compile (not sure if its possible to add the sybase jar).
The adapter uses jdbc to talk to tigrs chado database directly - its
read only. Tigrs database is not presently SO compliant, so this will
not yet work for fruitflys SO compliant chado - but thats the
challenge isnt it.

One contentious item is the chado-adapter.cfg, which is a database
configuration file, which is new ground for apollo. Its unclear if we
want separate database configs (theres alraedy a fair amount of config
files) or if we want to incorporate this into apollo.cfg (or will it
just clutter up apollo.cfg).